fix: tackle issue #1589
